driver_oncore.c: Report STATUS_TIME when device is reporting in position-hold mode.
It is not enough to have the poshold/acquiring flag, since this is
reported also in other circumstances with no fix (acquiring).
The device must also have reported to explicitly be in position hold mode.

test/daemon/oncore-binary-pos-hold.log.chk  updated since the device
now reports a fixed surveyed position.