driver_oncore.c: Poll the time mode (UTC/GPS).
Can be used to verify that time is set to be reported in UTC, for debug.

Note that before the device has received the leap seconds in the almanac
and thus knows the GPS-UTC offset, it reports time in GPS scale.  The Aw
message however also then says UTC, so the message cannot be used to
distinguish that.