build.sh: Make run on CentOS 7 and 8
CentOS 7 has Qt version 5, but it currently doesn't compile
with the default compiler, so stick with default Qt version.

CentOS 8 default Python version is 3, with "python" not being
defined by default, and scons binary being called "scons-3".

The additions were done such as to make only minimal changes
to existing logic to minimize potential for breaking that.
Tighter integration might be possible but was explicitly
avoided to ease independent maintenance for different
systems, as well as to allow easier backing out should that
be required/desired some time in the future.

ToDo: Have the script use "python3" for the Python shebang.
Explicitly setting the shebang that way isn't an issue when
calling scons(-3) interactively, but I've not found a good
way yet how to deal with the space embedded in the
python_shebang when calling scons(-3) from a shell script.
Suggestions welcome.