support gprmc sentences without faa mode indicator
For GPS receivers that emit GPRMC sentences without an FAA mode
indicator (pre-NMEA 2.3; for example Garmin GPS 16LVS which has NMEA
2.30 turned off by default), the `processRMC` [1] function should be
avoiding attempts to process a FAA mode indicator value. However, when
observing a valid sentence from a GPS device, the `STATUS_GPS` status is
not being set. For example, with the following sample data:

    $GPRMC,235954,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E*64
    $GPRMC,235955,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E*65
    $GPRMC,235956,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E*66
    $GPRMC,235957,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E*67
    $GPRMC,235958,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E*68
    $GPRMC,235959,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E*69

A user can observe a processed RMC sentence which reports a status value
of zero (ie. `STATUS_UNK`):

    $ gpsmon --nocurses -D 6
    ...
    gpsmon:DATA: NMEA0183: RMC: ddmmyy=071103 hhmmss=235958 lat=38.86 lon=-94.80 speed=0.00 track=221.90 mode=2 var=3.3 status=0
    ...

When inspecting how the count field is populated, it appears that the
discarding of the checksum will replace the asterisk with a comma [2],
which will result the empty checksum part to be counted as a field [3].
For example:

    (original)
    $GPRMC,235954,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E*64
    (discarded checksum)
    $GPRMC,235954,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E,
    (pre-count offset)
    GPRMC,235954,A,3851.3651,N,09447.9382,W,000.0,221.9,071103,003.3,E,
    (total count: 12)

With a count field of twelve [4], the RMC parsing assumes there is a FAA
mode indicator value (an empty field) which results in a `STATUS_UNK`
value (overriding a possibly previously determined `STATUS_GPS` or
`STATUS_DGPS` status).

This commit adjusts the count check to look for a field count greater
than twelve, which should cover the case of both legacy and newer NMEA
sentences. With this change, a desired `STATUS_GPS` status is detected:

    $ gpsmon --nocurses -D 6
    ...
    gpsmon:DATA: NMEA0183: RMC: ddmmyy=071103 hhmmss=235954 lat=38.86 lon=-94.80 speed=0.00 track=221.90 mode=2 var=3.3 status=1
    ...

[1]: https://gitlab.com/gpsd/gpsd/-/blob/a7a7bf12cbaa4c379655060c3ada1e80f1b8afa8/drivers/driver_nmea0183.c#L538
[2]: https://gitlab.com/gpsd/gpsd/-/blob/a7a7bf12cbaa4c379655060c3ada1e80f1b8afa8/drivers/driver_nmea0183.c#L4363-4370
[3]: https://gitlab.com/gpsd/gpsd/-/blob/a7a7bf12cbaa4c379655060c3ada1e80f1b8afa8/drivers/driver_nmea0183.c#L4383-4394
[4]: https://gitlab.com/gpsd/gpsd/-/blob/a7a7bf12cbaa4c379655060c3ada1e80f1b8afa8/drivers/driver_nmea0183.c#L673-674

Signed-off-by: James Knight <james.d.knight@live.com>