Don't reset the mode state timer if the mode wasn't reported in this epoch
Seen with an ATGM336H clone module, using the 杭州中科微电子有限公司 /
ZHONGKEWEI / icofchina.com AT6558R chip.  This chip emits NMEA and some
UBX messages.  It appears that it does not always emit messages with
the NMEA mode in every gpsd epoch.  cgps treats the missing mode as a
change in mode status, and resets the mode timer.  Thus cgps appears
to lose 3D fix every five or ten minutes, though the module is
actually using 12-15 satellites and holds a 3D fix for days.